\name{seedwater-package}
\alias{seedwater-package}
\alias{seedwater}
\docType{package}

\title{ Models for Drying and Soaking Kinetics of Seeds }

\description{
Bringing together tools for modeling drying and soaking (rehydration) 
kinetics of seeds. This package contains several widely used predictive models 
(e.g.: da Silva et al., 2018). As these are nonlinear, the functions are interactive-based 
and easy-to-use. Least squares estimates are obtained with just a few visual adjustments 
of the initial parameter values.
}

\details{
\tabular{ll}{
Package: \tab seedwater\cr
Type: \tab Package\cr
Version: \tab 2.1\cr
Date: \tab 2022-01-04\cr
License: \tab GPL (>= 2)\cr
}
}

\note{
\emph{seedwater} is an ongoing project.
Any and all criticism, comments and suggestions are welcomed.
}

\author{
Anderson Rodrigo da Silva

Maintainer: Anderson Rodrigo da Silva <anderson.silva@ifgoiano.edu.br>
}

\references{
da Silva, A. R. et al. (2018) Modeling the Three Phases of the Soaking Kinetics of Seeds. 
	\emph{Agronomy Journal}, 110:164-170. doi:10.2134/agronj2017.07.0373
}

\keyword{ package }
